backend/cortex/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
sqlalchemy
# LangChain Stack (LLM Logic)
langchain
langchain-community
langchain-ollama
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
# System
psutil
pynvml